+++ src/app/core/services/custom-date-adapter.service.ts
@@ -1,7 +1,6 @@
-import { Inject, Injectable, LOCALE_ID, PLATFORM_ID } from "@angular/core";
-import { NativeDateAdapter } from "@angular/material/core";
-import { Platform } from "@angular/cdk/platform";
-import { getLocaleFirstDayOfWeek } from "@angular/common";
+import { Inject, Injectable, LOCALE_ID, PLATFORM_ID } from '@angular/core';
+import { NativeDateAdapter } from '@angular/material/core';
+import { getLocaleFirstDayOfWeek } from '@angular/common';
 @Injectable()
 export class CustomDateAdapter extends NativeDateAdapter {
   constructor(
@@ -9,7 +8,6 @@ export class CustomDateAdapter extends NativeDateAdapter {
     @Inject(PLATFORM_ID) platformId
   ) {
     super(locale, platformId);
-    console.log(locale);
   }
   getFirstDayOfWeek() {
     // console.log("getFirstDayOfWeek()");